+++ generic/tclStringRep.h
@@ -47,15 +47,15 @@
  */
 
 typedef struct {
-    int numChars;		/* The number of chars in the string. -1 means
+    size_t numChars;		/* The number of chars in the string. -1 means
 				 * this value has not been calculated. >= 0
 				 * means that there is a valid Unicode rep, or
 				 * that the number of UTF bytes == the number
 				 * of chars. */
-    int allocated;		/* The amount of space actually allocated for
+    size_t allocated;		/* The amount of space actually allocated for
 				 * the UTF string (minus 1 byte for the
 				 * termination char). */
-    int maxChars;		/* Max number of chars that can fit in the
+    size_t maxChars;		/* Max number of chars that can fit in the
 				 * space allocated for the unicode array. */
     int hasUnicode;		/* Boolean determining whether the string has
 				 * a Unicode representation. */
